# specs/001-homelab-dashboard/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: HomeDash API
  version: 0.1.0
  description: |
    API contracts for HomeDash.

    Notes:
    - Cookie-based sessions are used for authentication.
    - CSRF protection is required for state-changing requests.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/admin/dashboards:
    get:
      tags: [admin]
      summary: List dashboards
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Dashboards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dashboard'
    post:
      tags: [admin]
      summary: Create dashboard
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
  /api/admin/dashboards/{dashboardId}:
    parameters:
      - name: dashboardId
        in: path
        required: true
        schema:
          type: string
    put:
      tags: [admin]
      summary: Update dashboard (name, applicability, background)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
    delete:
      tags: [admin]
      summary: Delete dashboard
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Deleted
  /api/admin/dashboards/{dashboardId}/layout:
    put:
      tags: [admin]
      summary: Replace dashboard layout (placeholders + widget instances)
      security:
        - cookieAuth: []
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardLayoutUpdateRequest'
      responses:
        '200':
          description: Updated
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: homedash_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    Dashboard:
      type: object
      required: [id, name, applicability]
      properties:
        id:
          type: string
        name:
          type: string
        applicability:
          type: string
          enum: [web, mobile, both]
        backgroundType:
          type: string
          enum: [solid, image]
          nullable: true
        backgroundColor:
          type: string
          nullable: true
        backgroundAssetId:
          type: string
          nullable: true
        backgroundDisplayMode:
          type: string
          enum: [fill, stretch]
          nullable: true
    DashboardCreateRequest:
      type: object
      required: [name, applicability]
      properties:
        name:
          type: string
        applicability:
          type: string
          enum: [web, mobile, both]
    DashboardUpdateRequest:
      type: object
      properties:
        name:
          type: string
        applicability:
          type: string
          enum: [web, mobile, both]
        backgroundType:
          type: string
          enum: [solid, image]
        backgroundColor:
          type: string
          nullable: true
        backgroundAssetId:
          type: string
          nullable: true
        backgroundDisplayMode:
          type: string
          enum: [fill, stretch]
          nullable: true
...
    DashboardLayoutUpdateRequest:
      type: object
      required: [placeholders]
      properties:
        placeholders:
          type: array
          items:
            $ref: '#/components/schemas/PlaceholderLayoutInput'
    PlaceholderLayoutInput:
      type: object
      required: [stableKey, x, y, w, h, borderColor, opacity, widgets]
      properties:
        stableKey:
          type: string
        title:
          type: string
          nullable: true
        x:
          type: integer
        y:
          type: integer
        w:
          type: integer
        h:
          type: integer
        borderColor:
          type: string
        opacity:
          type: number
        widgets:
          type: array
          items:
            $ref: '#/components/schemas/WidgetInstanceInput'
    WidgetInstanceInput:
      type: object
      required: [type, orderIndex, config]
      properties:
        type:
          type: string
        orderIndex:
          type: integer
        config:
          type: object
          additionalProperties: true